+++ data/HTMLBasedInterface/html_de/index.html
@@ -73,7 +73,7 @@
 					<li><a href="krebs.html?action=loadProject&amp;module=common_functions&amp;filename=projects/DNA.bvp">Brustkrebs</a></li>
 					<li><a href="laktose.html?action=loadProject&amp;module=common_functions&amp;filename=projects/lactase_1.bvp">Laktoseintoleranz</a></li>
 					<li><a href="ehec.html?action=loadProject&amp;module=common_functions&amp;filename=projects/intimin.bvp">EHEC</a></li>
-					<li><a href="wiki_tutorial.html?action=clearAll&amp;module=common_functions">BALLView Tutorial für Lehrer</a></li>
+					<li><a href="tutorial/wiki_tutorial01.html?action=clearAll&amp;module=common_functions">BALLView Tutorial für Lehrer</a></li>
 				</ul>
 			</div>
