+++ data/HTMLBasedInterface/html_de/tutorial/wiki_tutorial01.html
@@ -3,11 +3,11 @@
 	<head>
 		<title>Tutorial-Dialog</title>
 		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-		<link rel="stylesheet" href="../css/presentaball.css">
+		<link rel="stylesheet" href="../../css/presentaball.css">
 	</head>
 	<body>
 		<div class="navigation" id="breadcrumb">
-			<a href="index.html?action=clearAll&amp;module=common_functions">Startseite</a> > Wiki-Tutorial
+			<a href="../index.html?action=clearAll&amp;module=common_functions">Startseite</a> > Wiki-Tutorial
 		</div>
 
 		<div id="text-content">
@@ -18,7 +18,7 @@
 			<p>BALLView wurde ursprünglich für die wissenschaftliche Arbeit an der Universität im Bereich Bioinformatik programmiert. Hieraus wurde eine Schulversion für den Chemie- bzw. Biologieunterricht entwickelt. Mit BALLView können Lehrer und Schüler auf einfachste Weise Moleküle, wie Proteine oder DNA selbstständig bauen und dreidimensional drehen und betrachten. Aber auch größere Moleküle lassen sich durch die Anbindung verschiedener Datenbanken herunterladen und von allen Seiten betrachten.</p>
 
 			<ul class="buttons">
-				<li><a href="tutorial/wiki_tutorial02.html">weiter</a></li>
+				<li><a href="wiki_tutorial02.html">weiter</a></li>
 			</ul>
 		</div>
 	</body>
