+++ data/HTMLBasedInterface/html_de/tutorial/wiki_tutorial05.html
@@ -4,6 +4,7 @@
 		<title>Tutorial-Dialog</title>
 		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 		<link rel="stylesheet" href="../../css/presentaball.css">
+		<script src="qrc:///qtwebchannel/qwebchannel.js"></script>
 		<script>var current_step = 11;</script>
 		<script src="../../scripts/connectVIEW.js"></script>
 	</head>
