+++ data/HTMLBasedInterface/html_de/tutorial/wiki_tutorial08.html
@@ -5,6 +5,7 @@
 		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 		<link rel="stylesheet" href="../../css/presentaball.css">
 		<script>var current_step = 8;</script>
+		<script src="qrc:///qtwebchannel/qwebchannel.js"></script>
 		<script src="../../scripts/connectVIEW.js"></script>
 	</head>
 	<body>
