+++ data/HTMLBasedInterface/html_eng/index.html
@@ -71,7 +71,7 @@
 					<li><a href="imatinib.html?action=loadProject&amp;module=common_functions&amp;filename=projects/imatinib_complex.bvp">Leukemia</a></li>
 					<li><a href="hiv.html?action=loadProject&amp;module=common_functions&amp;filename=projects/2HMI_DNA.bvp">HIV</a></li>
 					<li><a href="ehec.html?action=loadProject&amp;module=common_functions&amp;filename=projects/intimin.bvp">EHEC</a></li>
-					<li><a href="tutorial_dialog.html?action=clearAll&amp;module=common_functions">BALLView demo tutorial</a></li>
+					<li><a href="tutorial/tutorial_dialog01.html?action=clearAll&amp;module=common_functions">BALLView demo tutorial</a></li>
 				</ul>
 			</div>
