+++ data/HTMLBasedInterface/html_eng/tutorial/tutorial_dialog01.html
@@ -3,13 +3,14 @@
 	<head>
 		<title>Tutorial dialog</title>
 		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-		<link rel="stylesheet" href="../css/presentaball.css">
+		<link rel="stylesheet" href="../../css/presentaball.css">
 		<script>var current_step = 1;</script>
-		<script src="../scripts/connectVIEW.js"></script>
+		<script src="qrc:///qtwebchannel/qwebchannel.js"></script>
+		<script src="../../scripts/connectVIEW.js"></script>
 	</head>
 	<body>
 		<div class="navigation" id="breadcrumb">
-			<a href="index.html?action=clearAll&amp;module=common_functions">Main page</a> > Tutorial dialog
+			<a href="../index.html?action=clearAll&amp;module=common_functions">Main page</a> > Tutorial dialog
 		</div>
 
 		<div id="text-content">
@@ -20,7 +21,7 @@
 			<p>Then enter the sequence ARND with the keyboard in the label field at the top (as an alternative you can also press the upper four amino acid buttons). This will create a polypeptide with the aminosequence (Alanin, Arginin, Asparagine, Aspartic acid).</p>
 			<p>Then press the <strong>Build</strong> button.This will enable the next step of the tutorial.</p>
 			<ul class="buttons">
-				<li><button id="button1" disabled onClick="location.href='tutorial/tutorial_dialog02.html';">Next</button></li>
+				<li><button id="button1" disabled onClick="location.href='tutorial_dialog02.html';">Next</button></li>
 			</ul>
 		</div>
 	</body>
