+++ data/HTMLBasedInterface/html_eng/tutorial/tutorial_dialog10.html
@@ -16,7 +16,7 @@
 			<p><strong>Congratulations</strong>, you successfully finished the the tutorial, which showed you a selection of BALLViews main features. Feel free to explore its other functions, e.g. by having a look at the BALLView documentation.</p>
 
 			<ul class="buttons">
-				<li><a href="../index.html">Main page</a></li>
+				<li><a href="../index.html?action=clearAll&amp;module=common_functions">Main page</a></li>
 			</ul>
 		</div>
 	</body>
