+++ source/EXTENSIONS/PRESENTABALL/CMakeLists.txt
@@ -5,8 +5,8 @@ IF(NOT BALL_HAS_VIEW)
 	RETURN()
 ENDIF()
 
-IF(NOT USE_QTWEBKIT OR NOT Qt5WebKit_FOUND)
-	MESSAGE(STATUS "QtWebkit support has not been enabled. Disabling the PresentaBALL plugin.")
+IF(NOT USE_QTWEBENGINE OR NOT Qt5WebEngine_FOUND)
+	MESSAGE(STATUS "QtWebEngine support has not been enabled. Disabling the PresentaBALL plugin.")
 	RETURN()
 ENDIF()
 
@@ -17,9 +17,10 @@ SET(UI_SOURCES_LIST
 )
 
 SET(SOURCES_LIST
+	${DIRECTORY}/PresentaBALLPage.C
 	${DIRECTORY}/PresentaBALLPlugin.C
 	${DIRECTORY}/PresentaBALLSettings.C
-	${DIRECTORY}/HTMLBasedInterface.C
+	${DIRECTORY}/PresentaBALLView.C
 )
 
 SET(RCC_SOURCES_LIST
@@ -29,9 +30,10 @@ SET(RCC_SOURCES_LIST
 SET(DIRECTORY include)
 ### the list of all files requiring a moc run ###
 SET(MOC_SOURCES_LIST
+	${DIRECTORY}/PresentaBALLPage.h
 	${DIRECTORY}/PresentaBALLPlugin.h
 	${DIRECTORY}/PresentaBALLSettings.h
-	${DIRECTORY}/HTMLBasedInterface.h
+	${DIRECTORY}/PresentaBALLView.h
 )
 
 QT5_WRAP_CPP(MOC_OUTFILES ${MOC_SOURCES_LIST})
@@ -63,4 +65,4 @@ INSTALL(TARGETS pluginPresentaBALL
 	LIBRARY DESTINATION "${BALL_PLUGIN_INSTALL_DIRECTORY}"
 )
 
-TARGET_LINK_LIBRARIES(pluginPresentaBALL BALL VIEW Qt5::WebKitWidgets)
+TARGET_LINK_LIBRARIES(pluginPresentaBALL BALL VIEW Qt5::WebEngineWidgets Qt5::WebChannel)
