+++ source/EXTENSIONS/PRESENTABALL/include/PresentaBALLPlugin.h
@@ -4,7 +4,7 @@
 #include <QtCore/QObject>
 #include <QtGui/QPixmap>
 
-#include <HTMLBasedInterface.h>
+#include <PresentaBALLView.h>
 #include <PresentaBALLSettings.h>
 
 #include <BALL/PLUGIN/BALLPlugin.h>
@@ -45,7 +45,7 @@ namespace BALL
 			private:
 				QPixmap icon_;
 				bool is_active_;
-				HTMLBasedInterface* html_interface_;
+				PresentaBALLView* html_interface_;
 				PresentaBALLSettings* settings_;
 		};
 	}
