+++ source/EXTENSIONS/PRESENTABALL/source/PresentaBALLSettings.C
@@ -1,5 +1,5 @@
 #include <PresentaBALLSettings.h>
-#include <HTMLBasedInterface.h>
+#include <PresentaBALLView.h>
 
 #include <QLineEdit>
 #include <QFileDialog>
@@ -10,7 +10,7 @@ namespace BALL
 {
 	namespace VIEW
 	{
-		PresentaBALLSettings::PresentaBALLSettings(HTMLBasedInterface* parent, const char* name, Qt::WindowFlags fl)
+		PresentaBALLSettings::PresentaBALLSettings(PresentaBALLView* parent, const char* name, Qt::WindowFlags fl)
 			: ConfigDialog(parent, fl),
 			  Ui_PresentaBALLSettingsData(),
 				html_interface_(parent)
